Add Write instances, use Err=Void for Buffers
