+++ Cargo.toml
@@ -2,3 +2,7 @@
 name = "coreio"
 version = "0.0.1"
 authors = []
+
+
+[dependencies]
+void = "*"
